+++ src/main/java/com/besson/arknights/block/custom/airtight/CleanroomCeilingLight.java
@@ -1,0 +1,24 @@
+package com.besson.arknights.block.custom.airtight;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModSwitchableLight;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+
+public class CleanroomCeilingLight extends ModSwitchableLight {
+
+    public CleanroomCeilingLight(Settings settings) {
+        super(settings.luminance(state -> state.get(LIT) ? 15 : 0));
+    }
+
+    private static final VoxelShape SHAPE = Block.createCuboidShape(4, 14, 4, 12, 16, 12);
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+}