+++ src/main/java/com/besson/arknights/block/custom/airtight/CleanroomDisinfectingQuarters.java
@@ -1,0 +1,32 @@
+package com.besson.arknights.block.custom.airtight;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModSwitchableLight;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.util.function.BooleanBiFunction;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.BlockView;
+
+import java.util.stream.Stream;
+
+public class CleanroomDisinfectingQuarters extends FurnitureHorizontalFacingBlock {
+
+    public CleanroomDisinfectingQuarters(Settings settings) {
+        super(settings);
+    }
+
+    private static final VoxelShape SHAPE_NS = Block.createCuboidShape(-2, 0, 0, 18, 32, 16);
+    private static final VoxelShape SHAPE_WE = Block.createCuboidShape(0, 0, -2, 16, 32, 18);
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return switch (state.get(FACING)){
+            case WEST,EAST -> SHAPE_WE;
+            default -> SHAPE_NS;
+        };
+    }
+}